Created a new buffer function for lw_geom geometries Minor code refactoring
