+++ geom/geom.c
@@ -7,87 +7,85 @@
 /*
 Parse WKT into GEOS geometry
 */
-GEOSGeometry* 
+GEOSGeometry *
 from_wkt(char *wkt)
 {
-    GEOSWKTReader *wkt_reader;
-    GEOSGeometry *geom;
+	GEOSWKTReader *wkt_reader;
+	GEOSGeometry *geom;
 
-    wkt_reader = GEOSWKTReader_create();
-    geom = GEOSWKTReader_read(wkt_reader, wkt);
+	wkt_reader = GEOSWKTReader_create();
+	geom = GEOSWKTReader_read(wkt_reader, wkt);
 
-    GEOSWKTReader_destroy(wkt_reader);  
+	GEOSWKTReader_destroy(wkt_reader);
 
-    return geom;  
+	return geom;
 }
 
 /*
 Get WKT from GEOS geometry
 */
-char*
+char *
 to_wkt(GEOSGeometry *g)
 {
-    GEOSWKTWriter *wkt_writer;
-    int output_dimention;
+	GEOSWKTWriter *wkt_writer;
+	int output_dimention;
 
-    wkt_writer = GEOSWKTWriter_create();
+	wkt_writer = GEOSWKTWriter_create();
 
-    // Set output dimention
+	// Set output dimention
 	output_dimention = GEOSGeom_getCoordinateDimension(g);
 	GEOSWKTWriter_setOutputDimension(wkt_writer, output_dimention);
 
-    char *wkt;
-    wkt = GEOSWKTWriter_write(wkt_writer, g);
+	char *wkt;
+	wkt = GEOSWKTWriter_write(wkt_writer, g);
 
-    GEOSWKTWriter_destroy(wkt_writer);
+	GEOSWKTWriter_destroy(wkt_writer);
 
-    return wkt;
+	return wkt;
 }
 
 /*
 Simplify a given geometry and generate a buffered geometry
 */
-GEOSGeometry*
+GEOSGeometry *
 simplified_buffer(GEOSGeometry *g, double width, double tolerance)
 {
-    GEOSGeometry *simple_geom;
-    GEOSGeometry *buffered_geom;
+	GEOSGeometry *simple_geom;
+	GEOSGeometry *buffered_geom;
 
-    // Simplify geometry with a given tolerance
-    simple_geom = GEOSTopologyPreserveSimplify(g, tolerance);
+	// Simplify geometry with a given tolerance
+	simple_geom = GEOSTopologyPreserveSimplify(g, tolerance);
 
-    //Create buffer
-    buffered_geom = GEOSBuffer(simple_geom, width, 8);
+	//Create buffer
+	buffered_geom = GEOSBuffer(simple_geom, width, 8);
 
-    //Memory cleanup
-    GEOSGeom_destroy(simple_geom);
+	//Memory cleanup
+	GEOSGeom_destroy(simple_geom);
 
-    return buffered_geom;
+	return buffered_geom;
 }
 
 /*
 Simplify and generate a buffer for a given WKT text.
 This function returns WKT.
 */
-char* 
+char *
 simplified_buffer_from_wkt(char *inwkt, double width, double tolerance)
 {
-    GEOSGeometry *in_geom;
-    GEOSGeometry *buffered_geom;
-    char *out_wkt;
+	GEOSGeometry *in_geom;
+	GEOSGeometry *buffered_geom;
+	char *out_wkt;
 
-    in_geom = from_wkt(inwkt);
-    buffered_geom = simplified_buffer(in_geom, width, tolerance);
+	in_geom = from_wkt(inwkt);
+	buffered_geom = simplified_buffer(in_geom, width, tolerance);
 
-    out_wkt = to_wkt(buffered_geom);
+	out_wkt = to_wkt(buffered_geom);
 
-    GEOSGeom_destroy(in_geom);
-    GEOSGeom_destroy(buffered_geom);
+	GEOSGeom_destroy(in_geom);
+	GEOSGeom_destroy(buffered_geom);
 
-    return out_wkt;
+	return out_wkt;
 }
-
-
 
 /*
 Return a linestring being a substring of the input one starting and ending at the given fractions of total 2d length.
@@ -95,58 +93,55 @@ Second and third arguments are float8 values between 0 and 1. This only works wi
 
 Copied from PostGIS source and modified accordingly.
 */
-LWGEOM*
+LWGEOM *
 lwgeom_line_substring(LWGEOM *ingeom, double from, double to)
 {
-	// GSERIALIZED *geom = PG_GETARG_GSERIALIZED_P(0);
-	// double from = PG_GETARG_FLOAT8(1);
-	// double to = PG_GETARG_FLOAT8(2);
+
 	LWGEOM *olwgeom = NULL;
 	POINTARRAY *ipa, *opa;
-	// GSERIALIZED *ret;
+
 	int type = lwgeom_get_type(ingeom);
 
-	// if ( from < 0 || from > 1 )
-	// {
-	// 	// elog(ERROR,"line_interpolate_point: 2nd arg isn't within [0,1]");
-	// 	return (int*)-1;
-	// }
+	if ( from < 0 || from > 1 )
+	{
+		// elog(ERROR,"line_interpolate_point: 2nd arg isn't within [0,1]");
+		return NULL;
+	}
 
-	// if ( to < 0 || to > 1 )
-	// {
-	// 	// elog(ERROR,"line_interpolate_point: 3rd arg isn't within [0,1]");
-	// 	return -1;
-	// }
+	if ( to < 0 || to > 1 )
+	{
+		// elog(ERROR,"line_interpolate_point: 3rd arg isn't within [0,1]");
+		return NULL;
+	}
 
-	// if ( from > to )
-	// {
-	// 	// elog(ERROR, "2nd arg must be smaller then 3rd arg");
-	// 	return -1;
-	// }
+	if ( from > to )
+	{
+		// elog(ERROR, "2nd arg must be smaller then 3rd arg");
+		return NULL;
+	}
 
-	if ( type == LINETYPE )
+	if (type == LINETYPE)
 	{
 		LWLINE *iline = lwgeom_as_lwline(ingeom);
 
-		if ( lwgeom_is_empty((LWGEOM*)iline) )
+		if (lwgeom_is_empty((LWGEOM *)iline))
 		{
 			/* TODO return empty line */
 			lwline_release(iline);
-			printf("test1");
-			// return -1;
+
+			return NULL;
 		}
 
 		ipa = iline->points;
 
 		opa = ptarray_substring(ipa, from, to, 0);
 
-		if ( opa->npoints == 1 ) /* Point returned */
+		if (opa->npoints == 1) /* Point returned */
 			olwgeom = (LWGEOM *)lwpoint_construct(iline->srid, NULL, opa);
 		else
 			olwgeom = (LWGEOM *)lwline_construct(iline->srid, NULL, opa);
-
 	}
-	else if ( type == MULTILINETYPE )
+	else if (type == MULTILINETYPE)
 	{
 		LWMLINE *iline;
 		uint32_t i = 0, g = 0;
@@ -156,31 +151,29 @@ lwgeom_line_substring(LWGEOM *ingeom, double from, double to)
 
 		iline = lwgeom_as_lwmline(ingeom);
 
-		if ( lwgeom_is_empty((LWGEOM*)iline) )
+		if (lwgeom_is_empty((LWGEOM *)iline))
 		{
 			/* TODO return empty collection */
 			lwmline_release(iline);
-			printf("test2");
-			
 		}
 
 		/* Calculate the total length of the mline */
-		for ( i = 0; i < iline->ngeoms; i++ )
+		for (i = 0; i < iline->ngeoms; i++)
 		{
-			LWLINE *subline = (LWLINE*)iline->geoms[i];
-			if ( subline->points && subline->points->npoints > 1 )
+			LWLINE *subline = (LWLINE *)iline->geoms[i];
+			if (subline->points && subline->points->npoints > 1)
 				length += ptarray_length_2d(subline->points);
 		}
 
-		geoms = lwalloc(sizeof(LWGEOM*) * iline->ngeoms);
+		geoms = lwalloc(sizeof(LWGEOM *) * iline->ngeoms);
 
 		/* Slice each sub-geometry of the multiline */
-		for ( i = 0; i < iline->ngeoms; i++ )
+		for (i = 0; i < iline->ngeoms; i++)
 		{
-			LWLINE *subline = (LWLINE*)iline->geoms[i];
+			LWLINE *subline = (LWLINE *)iline->geoms[i];
 			double subfrom = 0.0, subto = 0.0;
 
-			if ( subline->points && subline->points->npoints > 1 )
+			if (subline->points && subline->points->npoints > 1)
 				sublength += ptarray_length_2d(subline->points);
 
 			/* Calculate proportions for this subline */
@@ -189,25 +182,24 @@ lwgeom_line_substring(LWGEOM *ingeom, double from, double to)
 
 			/* This subline doesn't reach the lowest proportion requested
 			   or is beyond the highest proporton */
-			if ( from > maxprop || to < minprop )
+			if (from > maxprop || to < minprop)
 				continue;
 
-			if ( from <= minprop )
+			if (from <= minprop)
 				subfrom = 0.0;
-			if ( to >= maxprop )
+			if (to >= maxprop)
 				subto = 1.0;
 
-			if ( from > minprop && from <= maxprop )
+			if (from > minprop && from <= maxprop)
 				subfrom = (from - minprop) / (maxprop - minprop);
 
-			if ( to < maxprop && to >= minprop )
+			if (to < maxprop && to >= minprop)
 				subto = (to - minprop) / (maxprop - minprop);
 
-
 			opa = ptarray_substring(subline->points, subfrom, subto, 0);
-			if ( opa && opa->npoints > 0 )
+			if (opa && opa->npoints > 0)
 			{
-				if ( opa->npoints == 1 ) /* Point returned */
+				if (opa->npoints == 1) /* Point returned */
 				{
 					geoms[g] = (LWGEOM *)lwpoint_construct(SRID_UNKNOWN, NULL, opa);
 					homogeneous = LW_FALSE;
@@ -218,49 +210,81 @@ lwgeom_line_substring(LWGEOM *ingeom, double from, double to)
 				}
 				g++;
 			}
-
-
-
 		}
 		/* If we got any points, we need to return a GEOMETRYCOLLECTION */
-		if ( ! homogeneous )
+		if (!homogeneous)
 			type = COLLECTIONTYPE;
 
-		olwgeom = (LWGEOM*)lwcollection_construct(type, iline->srid, NULL, g, geoms);
+		olwgeom = (LWGEOM *)lwcollection_construct(type, iline->srid, NULL, g, geoms);
 	}
 	else
 	{
-		// elog(ERROR,"line_substring: 1st arg isn't a line");
-		// return -1;
-		printf("test3");
-
+		return NULL;
 	}
 
-    return olwgeom;
-
+	return olwgeom;
 }
 
-// LWGEOM*
-// geom_from_geojson(char *geojson)
-// {
-// 	LWGEOM *lwgeom;
-// 	char *srs = NULL;
+// Create a buffer around a geometry
+LWGEOM *
+buffer(LWGEOM *lwg, double width, int quadsegs)
+{
+	GEOSGeometry *geos_geom;
+	GEOSGeometry *buffered_geos;
+	LWGEOM *buffered_lwgeom;
 
-// 	lwgeom = lwgeom_from_geojson(geojson, &srs);
-// 	// if ( ! lwgeom )
-// 	// {
-// 	// 	/* Shouldn't get here */
-// 	// 	elog(ERROR, "lwgeom_from_geojson returned NULL");
-// 	// 	PG_RETURN_NULL();
-// 	// }
+	geos_geom = LWGEOM2GEOS(lwg, 0);
 
-// 	// if ( srs )
-// 	// {
-// 	// 	lwgeom_set_srid(lwgeom, getSRIDbySRS(srs));
-// 	// 	lwfree(srs);
-// 	// }
+	//create buffer
+	buffered_geos = GEOSBuffer(geos_geom, width, quadsegs);
 
-// 	// lwgeom_free(lwgeom);
-// 	return lwgeom;
+	if (!buffered_geos)
+		return NULL;
 
-// }
+	GEOSGeom_destroy(geos_geom);
+
+	if (lwgeom_has_srid(lwg))
+	{
+		GEOSSetSRID(buffered_geos, lwgeom_get_srid(lwg));
+	}
+
+	buffered_lwgeom = GEOS2LWGEOM(buffered_geos, 0);
+	if (!buffered_lwgeom)
+		return NULL;
+
+	GEOSGeom_destroy(buffered_geos);
+
+	return buffered_lwgeom;
+}
+
+// Create a buffer around a geometry using GEOSBufferParams
+LWGEOM *
+buffer_with_params(LWGEOM *lwg, double width, GEOSBufferParams *buffer_params)
+{
+	GEOSGeometry *geos_geom;
+	GEOSGeometry *buffered_geos;
+	LWGEOM *buffered_lwgeom;
+
+	geos_geom = LWGEOM2GEOS(lwg, 0);
+
+	//create buffer
+	buffered_geos = GEOSBufferWithParams(geos_geom, buffer_params, width);
+
+	if (!buffered_geos)
+		return NULL;
+
+	GEOSGeom_destroy(geos_geom);
+
+	if (lwgeom_has_srid(lwg))
+	{
+		GEOSSetSRID(buffered_geos, lwgeom_get_srid(lwg));
+	}
+
+	buffered_lwgeom = GEOS2LWGEOM(buffered_geos, 0);
+	if (!buffered_lwgeom)
+		return NULL;
+
+	GEOSGeom_destroy(buffered_geos);
+
+	return buffered_lwgeom;
+}
