+++ geom/geos.h
@@ -13,11 +13,6 @@
 void log_notice(const char *fmt, ...);
 void log_error(const char *fmt, ...);
 
-GEOSGeometry *from_wkt(char *wkt);
-char *to_wkt(GEOSGeometry *g);
-GEOSGeometry *simplified_buffer(GEOSGeometry *g, double width, double tolerance);
-char *simplified_buffer_from_wkt(char *inwkt, double width, double tolerance);
-
 GEOSContextHandle_t ctx;
 GEOSContextHandle_t init_geos_r();
 void init_geos();
