+++ geom/lwgeom.go
@@ -5,7 +5,7 @@ package geom
 #include <geos_c.h>
 #include "lwgeom_geos.h"
 #include "geos.h"
-#include "lwgeom.h"
+#include "geom.h"
 char *cnull = NULL;
 */
 import (
@@ -16,40 +16,60 @@ import (
 	"unsafe"
 )
 
-// LwGeom Go type to wrap lwgeom
-type LwGeom struct {
+// Geom Go type to wrap lwgeom
+type Geom struct {
 	LwGeom *C.LWGEOM
 }
 
-// LwGeomFromGeoJSON creates lwgeom from GeoJson
-func LwGeomFromGeoJSON(geojson string) *LwGeom {
+// SRS Stores required spatial reference systems
+var SRS = map[string]string{
+	"EPSG:4326": "+proj=longlat +datum=WGS84 +no_defs",
+	"EPSG:3857": "+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext  +no_defs",
+}
+
+// LwGeomVersion the version number of liblwgeom
+func LwGeomVersion() string {
+	version := C.lwgeom_version()
+	
+	return C.GoString(version)
+}
+
+// GEOSVersion returns the version number of libgeos
+func GEOSVersion() string {
+	geosVersion := C.lwgeom_geos_version()
+
+	return C.GoString(geosVersion)
+}
+
+// FromGeoJSON creates lwgeom from GeoJson
+func FromGeoJSON(geojson string) *Geom {
 
 	geojsonCstring := C.CString(geojson)
 	lwgeom := C.lwgeom_from_geojson(geojsonCstring, &C.cnull)
 	defer C.lwfree(unsafe.Pointer(geojsonCstring))
 
-	return &LwGeom{
+	return &Geom{
 		LwGeom: lwgeom,
 	}
 }
 
 // LwGeomFromGEOS convert GEOS geometry to lwgeom
-func LwGeomFromGEOS(geosGeom *C.GEOSGeometry) *LwGeom {
+func LwGeomFromGEOS(geosGeom *C.GEOSGeometry) *Geom {
 
 	lwGeom := C.GEOS2LWGEOM(geosGeom, C.uchar(0))
 
-	return &LwGeom{
+	return &Geom{
 		LwGeom: lwGeom,
 	}
 }
 
 // Free clears the memory allocated to lwgeom
-func (lwg *LwGeom) Free() {
+func (lwg *Geom) Free() {
 	C.lwgeom_free(lwg.LwGeom)
 }
 
 // ToGeoJSON generates geojson from lwgeom
-func (lwg *LwGeom) ToGeoJSON(precisoin int, hasBbox int) string {
+func (lwg *Geom) ToGeoJSON(precisoin int, hasBbox int) string {
 
 	geojson := C.lwgeom_to_geojson(lwg.LwGeom, C.cnull, C.int(precisoin), C.int(hasBbox))
 	defer C.lwfree(unsafe.Pointer(geojson))
@@ -57,7 +77,7 @@ func (lwg *LwGeom) ToGeoJSON(precisoin int, hasBbox int) string {
 }
 
 // LineSubstring returns a part of the linestring
-func (lwg *LwGeom) LineSubstring(from float64, to float64) {
+func (lwg *Geom) LineSubstring(from float64, to float64) {
 
 	defer C.lwgeom_free(lwg.LwGeom)
 
@@ -66,17 +86,17 @@ func (lwg *LwGeom) LineSubstring(from float64, to float64) {
 }
 
 // SetSRID sets the SRID of the geometry
-func (lwg *LwGeom) SetSRID(srid int) {
+func (lwg *Geom) SetSRID(srid int) {
 	C.lwgeom_set_srid(lwg.LwGeom, C.int(srid))
 }
 
 // GetSRID returns the SRID of the geometry
-func (lwg *LwGeom) GetSRID() int {
+func (lwg *Geom) GetSRID() int {
 	return int(C.lwgeom_get_srid(lwg.LwGeom))
 }
 
 // ToGEOS converts lwgeom to GEOS geometry
-func (lwg *LwGeom) ToGEOS() *GEOSGeom {
+func (lwg *Geom) ToGEOS() *GEOSGeom {
 
 	return GenerateGeosGeom(C.LWGEOM2GEOS(lwg.LwGeom, C.uchar(0)))
 }
@@ -86,7 +106,7 @@ Project transforms (reproject) a geometry from one SRS to another.
 You will have to use the WKT versions of SRS definition.
 references: https://epsg.io
 */
-func (lwg *LwGeom) Project(fromSRS string, toSRS string) {
+func (lwg *Geom) Project(fromSRS string, toSRS string) {
 	fromSrsCtype := C.CString(fromSRS)
 	defer C.free(unsafe.Pointer(fromSrsCtype))
