+++ geom/lwgeom_test.go
@@ -6,7 +6,7 @@ import (
 
 func TestGeomFromGeoJson(t *testing.T) {
 
-	geom := LwGeomFromGeoJSON(JSONLinestring)
+	geom := FromGeoJSON(JSONLinestring)
 
 	if geom == nil {
 		t.Error("Error: GeomFromGeoJson()")
@@ -17,7 +17,7 @@ func TestGeomFromGeoJson(t *testing.T) {
 
 func TestToGeoJson(t *testing.T) {
 
-	geom := LwGeomFromGeoJSON(JSONLinestring)
+	geom := FromGeoJSON(JSONLinestring)
 	jsonString := geom.ToGeoJSON(4, 0)
 
 	if jsonString == "" {
@@ -31,7 +31,7 @@ func TestLineSubstring(t *testing.T) {
 
 	expectedJSON := `{"type":"LineString","coordinates":[[79.9066,6.8597],[79.9073,6.859],[79.9076,6.8588],[79.9078,6.8585],[79.908,6.8582],[79.9083,6.858],[79.9084,6.8579],[79.9085,6.8578],[79.9088,6.8575],[79.9089,6.8573],[79.9089,6.8573]]}`
 
-	geom := LwGeomFromGeoJSON(JSONLinestring)
+	geom := FromGeoJSON(JSONLinestring)
 	geom.LineSubstring(0.5, 0.52)
 	resultJSON := geom.ToGeoJSON(4, 0)
 
@@ -44,7 +44,7 @@ func TestLineSubstring(t *testing.T) {
 
 func TestToGEOS(t *testing.T) {
 
-	geom := LwGeomFromGeoJSON(JSONLinestring)
+	geom := FromGeoJSON(JSONLinestring)
 	geos := geom.ToGEOS()
 
 	coords, _ := geos.GetNumCoordinates()
@@ -74,7 +74,7 @@ func TestLwGeomFromGEOS(t *testing.T) {
 
 func TestProject(t *testing.T) {
 
-	geom := LwGeomFromGeoJSON(JSONLinestring)
+	geom := FromGeoJSON(JSONLinestring)
 	geom.SetSRID(4326)
 
 	fromSRS := SRS["EPSG:4326"]
@@ -98,4 +98,20 @@ func TestProject(t *testing.T) {
 
 	geom.Free()
 	geos.Destroy()
+}
+
+func TestLwGeomVersion(t *testing.T) {
+	version := LwGeomVersion()
+
+	if version == "" {
+		t.Error("Error: LwGeomVersion()")
+	}
+}
+
+func TestGEOSVersion(t *testing.T) {
+	version := GEOSVersion()
+
+	if version == "" {
+		t.Error("Error: GEOSVersion()")
+	}
 }
